requirement.txt
aioredis==2.0.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.1
fastapi-mail==1.0.8
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iso8601==1.0.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pypika-tortoise==0.1.4
python-multipart==0.0.5
pytz==2022.1
redis==4.1.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.17.1
tortoise-orm==0.19.0
typing_extensions==4.1.1
uvicorn==0.17.6
wrapt==1.14.0
